feat: фикс багов

+++ src/migrations/001_listings_migration.sql
@@ -11,7 +11,7 @@ CREATE TABLE listings (
     phone VARCHAR(255),
     title VARCHAR(255) NOT NULL,
     description TEXT,
-    price DECIMAL(10, 2) NOT NULL,
+    price INT NOT NULL,
     images TEXT[] DEFAULT ARRAY[]::TEXT[], 
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
